+++ x/gov/genesis_test.go
@@ -3,9 +3,8 @@ package gov
 import (
 	"testing"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
-
-	"github.com/cosmos/cosmos-sdk/x/mock"
 
 	abci "github.com/tendermint/tendermint/abci/types"
 )
@@ -78,7 +77,7 @@ func TestImportExportQueues(t *testing.T) {
 	require.True(t, keeper.GetProposal(ctx, proposalID1).GetStatus() == StatusDepositPeriod)
 	require.True(t, keeper.GetProposal(ctx, proposalID2).GetStatus() == StatusVotingPeriod)
 
-	genAccs := mock.GetAllAccounts(mapp.AccountKeeper, ctx)
+	genAccs := sdk.GetAllAccounts(mapp.AccountKeeper, ctx)
 
 	// Export the state and import it into a new Mock App
 	genState := ExportGenesis(ctx, keeper)