+++ x/staking/simulation/invariants.go
@@ -6,21 +6,18 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-	"github.com/cosmos/cosmos-sdk/x/distribution"
-	"github.com/cosmos/cosmos-sdk/x/mock/simulation"
 	"github.com/cosmos/cosmos-sdk/x/staking"
 	"github.com/cosmos/cosmos-sdk/x/staking/keeper"
 )
 
 // AllInvariants runs all invariants of the staking module.
 // Currently: total supply, positive power
-func AllInvariants(ck bank.Keeper, k staking.Keeper,
-	f auth.FeeCollectionKeeper, d distribution.Keeper,
-	am auth.AccountKeeper) simulation.Invariant {
+func AllInvariants(k staking.Keeper,
+	f FeeCollectionKeeper, d DistributionKeeper,
+	am auth.AccountKeeper) sdk.Invariant {
 
 	return func(ctx sdk.Context) error {
-		err := SupplyInvariants(ck, k, f, d, am)(ctx)
+		err := SupplyInvariants(k, f, d, am)(ctx)
 		if err != nil {
 			return err
 		}
@@ -46,8 +43,8 @@ func AllInvariants(ck bank.Keeper, k staking.Keeper,
 
 // SupplyInvariants checks that the total supply reflects all held not-bonded tokens, bonded tokens, and unbonding delegations
 // nolint: unparam
-func SupplyInvariants(ck bank.Keeper, k staking.Keeper,
-	f auth.FeeCollectionKeeper, d distribution.Keeper, am auth.AccountKeeper) simulation.Invariant {
+func SupplyInvariants(k staking.Keeper,
+	f auth.FeeCollectionKeeper, d DistributionKeeper, am auth.AccountKeeper) sdk.Invariant {
 	return func(ctx sdk.Context) error {
 		pool := k.GetPool(ctx)
 
@@ -73,16 +70,14 @@ func SupplyInvariants(ck bank.Keeper, k staking.Keeper,
 			return false
 		})
 
-		feePool := d.GetFeePool(ctx)
-
 		// add outstanding fees
 		loose = loose.Add(sdk.NewDecFromInt(f.GetCollectedFees(ctx).AmountOf(k.BondDenom(ctx))))
 
 		// add community pool
-		loose = loose.Add(feePool.CommunityPool.AmountOf(k.BondDenom(ctx)))
+		loose = loose.Add(d.GetFeePoolCommunityCoins(ctx).AmountOf(k.BondDenom(ctx)))
 
 		// add yet-to-be-withdrawn
-		loose = loose.Add(d.GetOutstandingRewards(ctx).AmountOf(k.BondDenom(ctx)))
+		loose = loose.Add(d.GetOutstandingRewardsCoins(ctx).AmountOf(k.BondDenom(ctx)))
 
 		// Not-bonded tokens should equal coin supply plus unbonding delegations
 		// plus tokens on unbonded validators
@@ -104,7 +99,7 @@ func SupplyInvariants(ck bank.Keeper, k staking.Keeper,
 }
 
 // NonNegativePowerInvariant checks that all stored validators have >= 0 power.
-func NonNegativePowerInvariant(k staking.Keeper) simulation.Invariant {
+func NonNegativePowerInvariant(k staking.Keeper) sdk.Invariant {
 	return func(ctx sdk.Context) error {
 		iterator := k.ValidatorsPowerStoreIterator(ctx)
 
@@ -132,7 +127,7 @@ func NonNegativePowerInvariant(k staking.Keeper) simulation.Invariant {
 }
 
 // PositiveDelegationInvariant checks that all stored delegations have > 0 shares.
-func PositiveDelegationInvariant(k staking.Keeper) simulation.Invariant {
+func PositiveDelegationInvariant(k staking.Keeper) sdk.Invariant {
 	return func(ctx sdk.Context) error {
 		delegations := k.GetAllDelegations(ctx)
 		for _, delegation := range delegations {
@@ -151,7 +146,7 @@ func PositiveDelegationInvariant(k staking.Keeper) simulation.Invariant {
 // DelegatorSharesInvariant checks whether all the delegator shares which persist
 // in the delegator object add up to the correct total delegator shares
 // amount stored in each validator
-func DelegatorSharesInvariant(k staking.Keeper) simulation.Invariant {
+func DelegatorSharesInvariant(k staking.Keeper) sdk.Invariant {
 	return func(ctx sdk.Context) error {
 		validators := k.GetAllValidators(ctx)
 		for _, validator := range validators {