+++ x/staking/types/validator.go
@@ -3,7 +3,6 @@ package types
 import (
 	"bytes"
 	"fmt"
-	"math/big"
 	"strings"
 	"time"
 
@@ -417,23 +416,9 @@ func (v Validator) TendermintPower() int64 {
 	return 0
 }
 
-var powerReduction = sdk.NewIntFromBigInt(new(big.Int).Exp(big.NewInt(10), big.NewInt(6), nil))
-
 // potential Tendermint power
 func (v Validator) PotentialTendermintPower() int64 {
-	return (v.Tokens.Div(powerReduction)).Int64()
-}
-
-// utility functions
-
-// TokensToTendermintPower - convert input tokens to potential tendermint power
-func TokensToTendermintPower(tokens sdk.Int) int64 {
-	return (tokens.Div(powerReduction)).Int64()
-}
-
-// TokensFromTendermintPower - convert input power to tokens
-func TokensFromTendermintPower(power int64) sdk.Int {
-	return sdk.NewInt(power).Mul(powerReduction)
+	return sdk.TokensToTendermintPower(v.Tokens)
 }
 
 // ensure fulfills the sdk validator types